Fix packets page column mapping and implement clean URL navigation
- Fix packets table to show projectName instead of non-existent packetName
- Restructure navigation to use clean URLs: /packets, /packets/pieces, /packets/presets
- Create shared PacketsNavigation component with consistent tab navigation
- Update all edit forms to redirect to correct section after save/cancel
- Each section now has its own dedicated page with proper data loading
- Maintain tab navigation across all pages for easy switching between sections
